+++ client/src/features/myCollections/EditIdentifiersForm.tsx
@@ -152,8 +152,8 @@ const EditIdentifiersFrom: React.FC<IProps> = ({ data, onClose }) => {
         <Col className="d-flex justify-content-start">
           <SecondaryButton
             type="button"
-            textColor={theme.color.button}
-            actionBgColor={theme.color.button}
+            $textColor={theme.color.button}
+            $actionBgColor={theme.color.button}
             onClick={() => handleAddIdentifierRow()}
           >
             <i className="bi bi-plus-lg mx-2 d-inline-block ms-0" />