+++ client/src/test/data/concept/items.ts
@@ -1,0 +1,59 @@
+export const items = {
+  id: 'https://lux.collections.yale.edu/data/concept/b34ad31c-4a2f-4cef-aee3-ae8d824e42a2',
+  type: 'Type',
+  _label: 'items',
+  equivalent: [
+    {
+      id: 'http://vocab.getty.edu/aat/300404024',
+      type: 'Type',
+      _label: 'items',
+    },
+  ],
+  identified_by: [
+    {
+      type: 'Name',
+      content: 'items',
+      language: [
+        {
+          id: 'https://lux.collections.yale.edu/data/concept/dfa53b96-4eda-4c9a-b091-10008a726c38',
+          type: 'Language',
+          _label: 'English',
+          notation: 'en',
+          equivalent: [
+            {
+              id: 'http://vocab.getty.edu/aat/300388277',
+              type: 'Language',
+              _label: 'English',
+            },
+          ],
+        },
+      ],
+      classified_as: [
+        {
+          id: 'https://lux.collections.yale.edu/data/concept/f7ef5bb4-e7fb-443d-9c6b-371a23e717ec',
+          type: 'Type',
+          _label: 'Primary Name',
+          equivalent: [
+            {
+              id: 'http://vocab.getty.edu/aat/300404670',
+              type: 'Type',
+              _label: 'Primary Name',
+            },
+          ],
+        },
+        {
+          id: 'https://lux.collections.yale.edu/data/concept/ab99d278-9323-4d84-8e97-1846058fc587',
+          type: 'Type',
+          _label: 'Sort Name',
+          equivalent: [
+            {
+              id: 'http://vocab.getty.edu/aat/300451544',
+              type: 'Type',
+              _label: 'Final: Sort Title',
+            },
+          ],
+        },
+      ],
+    },
+  ],
+}