+++ client/src/test/data/concept/preferredTerms.ts
@@ -1,0 +1,12 @@
+export const preferredTerms = {
+  id: 'https://lux.collections.yale.edu/data/concept/f7ef5bb4-e7fb-443d-9c6b-371a23e717ec',
+  type: 'Type',
+  _label: 'preferred terms',
+  equivalent: [
+    {
+      id: 'http://vocab.getty.edu/aat/300404670',
+      type: 'Type',
+      _label: 'preferred terms',
+    },
+  ],
+}