+++ client/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Changed
+
+- Allowed iframe tags in CMS contents ([#524](https://github.com/project-lux/lux-frontend/issues/524)).
+
+### Fixed
+
+- Fixed hooks being incorrectly used with ApiText ([#808](https://github.com/project-lux/lux-frontend/issues/808)).
+- Fixed advanced search config not being retrieved from non-home landing pages ([#818](https://github.com/project-lux/lux-frontend/issues/818)).
+
 ## v1.60.2 - 2025-11/20
 
 ### Fixed